PHM_ADMIN_TOOLS/LOOKER_EMAIL_SCHEDULERS/SRC/requirements.txt
looker-sdk==24.8.0
sendgrid==6.11.0
pathvalidate==3.2.0
google-cloud-bigquery==3.24.0
google-cloud-storage==2.16.0
pandas==2.2.2
google-auth==2.29.0
typer==0.12.3
python-dotenv==1.0.1
